feat!: support installing multiple binaries
Please note the `repo` input has been renamed to `targets` to more
clearly describe the input's behavior in light of this new feature.

To install multiple binaries, pass a whitespace-separated list:

```yaml
- name: Install flux-capacitor
  uses: EricCrosson/install-github-release-binary@v1
  with:
    targets: |
      EricCrosson/flux-capacitor@v1
      EricCrosson/steam-locomotive@v7.5.3
```

Why the name `targets` and not `target`? Or maybe an alias from
`target` to `targets`?

If we accept multiple inputs, one of which has to be present,
the simple input metadata can no longer describe this contract.

Choose `targets` over `target` because a list of one is still a list!

BREAKING CHANGE: rename `repo` to `targets`