fix: remove support for specifying targets based on sha
Closes #26

This feature was an attempt at reprodubitility but did not actually
end up helping. I'm removing this feature to reduce the API surface
area to foster maintainability and to remove a potential foot gun.

I'm not publishing this as a breaking change because it was added so
recently I do not expect many users, and I want to be able to continue
pushing updates to v2 users without the overhead of multiple release
branches. I do not expect this to impact many (any?) users. If this
does affect you, feel free to reach out by opening a ticket!

Famous last words? Stay tuned!